+++ islands/EntryInput/index.tsx
@@ -1,9 +1,8 @@
-import { useEffect, useRef } from 'preact/hooks';
-import { Chiplist, Text } from 'lunchbox';
+import { Card, Chiplist, Text } from 'lunchbox';
 import IconTag from 'icons/tag.tsx';
 import IconBookmark from 'icons/bookmark.tsx';
-import Handlers from '@/utils/handlers/EntryInput.ts';
 import { iEntry } from '@/utils/db/entry.ts';
+import { iQueryTags, iTag } from '@/utils/db/tag.ts';
 import EntryTypeIndicator from '@/components/EntryTypeIndicator/index.tsx';
 import EntryLengthIndicator from '@/components/EntryLengthIndicator/index.tsx';
 import {
@@ -11,78 +10,95 @@ import {
   ENTRY_INPUT_FIELD,
   ICON_STANDARD,
 } from '@/utils/styles.ts';
+import InputContent from './components/InputContent.tsx';
+import { createEntryInputState } from '@/utils/hooks.ts';
 
 export interface iEntryInput {
   entry: iEntry;
-  entryId: string;
-  onFocusOut: () => void;
+  userTags: iTag[];
+  isLocal?: boolean;
 }
 
 export default function EntryInput(props: iEntryInput) {
-  const refTextarea = useRef<HTMLSpanElement>(null);
+  const state = createEntryInputState(
+    props.userTags,
+    props.entry,
+  );
 
-  const {
-    handleRemoveTag,
-    handleEntryMarkInput,
-    handleTagInput,
-    handleEntryInput,
-    handleFieldFocus,
-    handleConatinerKeyDown,
-    entryMark,
-    tagIds,
-    entryValue,
-  } = Handlers(props);
-
-  useEffect(() => {
-    refTextarea.current?.focus();
-    if (entryValue === '' && refTextarea.current) {
-      refTextarea.current.innerHTML = '';
-    }
-  }, [entryValue]);
+  // References
+  const entryTags = state.userTags.filter((userTag) =>
+    state.entry.tagIds.includes(userTag.id)
+  );
 
   return (
-    <div>
-      <div class='isl-EntryInput-container' onKeyDown={handleConatinerKeyDown}>
+    <Card>
+      <div class='isl-EntryInput-container'>
         <div class={ENTRY_GRID}>
-          {/* Textarea Row */}
-          <EntryTypeIndicator tags={tagIds} />
-          <Text
-            contentEditable
-            onKeyUp={handleEntryInput}
-            fref={refTextarea}
-            class={`${ENTRY_INPUT_FIELD} px-1.5`}
-            style={{ lineBreak: 'anywhere' }}
-          >
-            {props.entry.content}
-          </Text>
-
-          {/* Entry Mark Row */}
+          <EntryTypeIndicator
+            tags={entryTags.map((entryTag) => entryTag.name)}
+          />
+          <InputContent {...state} />
+          {
+            /*
           <IconBookmark class={ICON_STANDARD} />
           <input
             class={`${ENTRY_INPUT_FIELD} px-1.5`}
             type='text'
-            onKeyUp={handleEntryMarkInput}
-            onFocus={handleFieldFocus('entrymark')}
-            value={entryMark}
-          />
-
-          {/* Tags Row */}
-          <IconTag class={ICON_STANDARD} />
-          <input
-            class={`${ENTRY_INPUT_FIELD} px-1.5`}
-            type='text'
-            onFocus={handleFieldFocus('tags')}
-            onKeyUp={handleTagInput}
-          />
+            // onKeyUp={handleEntryMarkInput}
+            // onFocus={handleFieldFocus('entrymark')}
+            // value={entryMark}
+          /> */
+          }
         </div>
-        {/* Tags Chiplist */}
-        <Chiplist
-          values={tagIds}
-          onRemove={handleRemoveTag}
-          class='ml-6 mt-1.5'
-        />
       </div>
-      <EntryLengthIndicator length={entryValue.length} />
-    </div>
+      <EntryLengthIndicator length={0} />
+    </Card>
   );
+
+  // return (
+  //   <div>
+  //     <div class='isl-EntryInput-container' onKeyDown={handleConatinerKeyDown}>
+  //       <div class={ENTRY_GRID}>
+  //         {/* Textarea Row */}
+  //         <EntryTypeIndicator tags={tagIds} />
+  //         <Text
+  //           contentEditable
+  //           onKeyUp={handleEntryInput}
+  //           fref={refTextarea}
+  //           class={`${ENTRY_INPUT_FIELD} px-1.5`}
+  //           style={{ lineBreak: 'anywhere' }}
+  //         >
+  //           {props.entry.content}
+  //         </Text>
+
+  //         {/* Entry Mark Row */}
+  //         <IconBookmark class={ICON_STANDARD} />
+  //         <input
+  //           class={`${ENTRY_INPUT_FIELD} px-1.5`}
+  //           type='text'
+  //           onKeyUp={handleEntryMarkInput}
+  //           onFocus={handleFieldFocus('entrymark')}
+  //           value={entryMark}
+  //         />
+
+  //         {/* Tags Row */}
+  //         <IconTag class={ICON_STANDARD} />
+  //         <input
+  //           class={`${ENTRY_INPUT_FIELD} px-1.5`}
+  //           type='text'
+  //           onFocus={handleFieldFocus('tags')}
+  //           onKeyUp={handleTagInput}
+  //         />
+  //       </div>
+  //       {/* Tags Chiplist */}
+  //       <Chiplist
+  //         values={tagIds}
+  //         onRemove={handleRemoveTag}
+  //         class='ml-6 mt-1.5'
+  //       />
+  //     </div>
+  //     <EntryLengthIndicator length={entryValue.length} />
+
+  //   </div>
+  // );
 }