+++ utils/db/tag.ts
@@ -1,4 +1,5 @@
 export interface iTag {
+  id: string;
   name: string;
 }